(355) Refactor utilities for budget domain
